requirements.txt
wheel>=0.30.0
prettytable>=0.7.2
PyYAML>=3.12
argparse>=1.1
colour_demosaicing>=0.1.4
opencv-python>=4.1.0
imageio>=2.5.0
joblib>=0.13.2
matplotlib>=3.1.0
numpy>=1.17.3
pandas>=0.23.4
pillow>=5.1.0
plotly>=3.6.1
plyfile>=0.7.2
prettytable>=0.7.2
pynmea2>=1.15.0
pytz>=2018.9
recommonmark>=0.6.0
scipy>=1.2.1
sphinx_rtd_theme>=0.4.3
tqdm>=4.32.2
scikit-image>=0.17.2

setuptools~=42.0.2
numba~=0.51.2